Move all model-specific code to plugin subdirectories
