+++ setup.cfg
@@ -59,12 +59,12 @@ zip_safe = False
 
 [options.entry_points]
 ewatercycle.models =
-    Hype = ewatercycle.models.hype:Hype
-    Lisflood = ewatercycle.models.lisflood:Lisflood
-    Wflow = ewatercycle.models.wflow:Wflow
-    PCRGlobWB = ewatercycle.models.pcrglobwb:PCRGlobWB
-    MarrmotM01 = ewatercycle.models.marrmot:MarrmotM01
-    MarrmotM14 = ewatercycle.models.marrmot:MarrmotM14
+    Hype = ewatercycle.plugins.hype.ewc_model:Hype
+    Lisflood = ewatercycle.plugins.lisflood.ewc_model:Lisflood
+    Wflow = ewatercycle.plugins.wflow.ewc_model:Wflow
+    PCRGlobWB = ewatercycle.plugins.pcrglobwb.ewc_model:PCRGlobWB
+    MarrmotM01 = ewatercycle.plugins.marrmot.ewc_model:MarrmotM01
+    MarrmotM14 = ewatercycle.plugins.marrmot.ewc_model:MarrmotM14
 
 [options.packages.find]
 where = src
