+++ src/ewatercycle/plugins/pcrglobwb/_pcrglobwb.py
@@ -5,14 +5,14 @@ from typing import Optional
 
 from esmvalcore.experimental import get_recipe
 
-from ..util import (
+from ...util import (
     data_files_from_recipe_output,
     get_extents,
     get_time,
     to_absolute_path,
 )
-from ._default import DefaultForcing, _session
-from .datasets import DATASETS
+from ...forcing._default import DefaultForcing, _session
+from ...forcing.datasets import DATASETS
 
 
 class PCRGlobWBForcing(DefaultForcing):
