+++ src/ewatercycle/plugins/pcrglobwb/ewc_model.py
@@ -13,7 +13,7 @@ from grpc4bmi.bmi_client_docker import BmiClientDocker
 from grpc4bmi.bmi_client_singularity import BmiClientSingularity
 
 from ewatercycle import CFG
-from ewatercycle.forcing._pcrglobwb import PCRGlobWBForcing
+from ewatercycle.plugins.pcrglobwb._pcrglobwb import PCRGlobWBForcing
 from ewatercycle.models.abstract import AbstractModel
 from ewatercycle.parameter_sets import ParameterSet
 from ewatercycle.parametersetdb.config import CaseConfigParser
