+++ src/ewatercycle/plugins/wflow/forcing.py
@@ -3,9 +3,9 @@ from typing import Dict, Optional
 
 from esmvalcore.experimental import get_recipe
 
-from ..util import get_extents, get_time, to_absolute_path
-from ._default import DefaultForcing, _session
-from .datasets import DATASETS
+from ...util import get_extents, get_time, to_absolute_path
+from ...forcing._default import DefaultForcing, _session
+from ...forcing.datasets import DATASETS
 
 
 class WflowForcing(DefaultForcing):
