+++ tests/models/test_marrmotm01.py
@@ -10,7 +10,7 @@ from xarray.testing import assert_allclose
 
 from ewatercycle import CFG
 from ewatercycle.forcing import load_foreign
-from ewatercycle.models.marrmot import MarrmotM01, Solver
+from ewatercycle.plugins.marrmot.ewc_model import MarrmotM01, Solver
 
 
 @pytest.fixture
